requirements.txt
# Core dependencies (původní)
pandas
openpyxl

# Azure Functions dependencies
azure-functions